requirements.txt
flask
requests
python-dotenv
pymongo[srv]
alpaca-py
nest-asyncio